services/user-service/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic[email]==2.5.0
pydantic-settings==2.1.0
boto3==1.34.0
python-jose[cryptography]==3.3.0
python-multipart==0.0.6
PyJWT==2.8.0
cryptography==41.0.7
requests==2.31.0
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.2
moto[dynamodb]==4.2.9
